+++ variant/src/lut.cc
@@ -34,7 +34,7 @@ void LUTBase<inputs_2_merged_LUT_is_input_ordered, merged_2_inputs_LUT_is_input_
 template<bool inputs_2_merged_LUT_is_input_ordered, bool merged_2_inputs_LUT_is_input_ordered>
 void LUTBase<inputs_2_merged_LUT_is_input_ordered, merged_2_inputs_LUT_is_input_ordered>::reset_vector(vector<int64_t>& vec, int64_t from)
 {
-  for(auto i=from;i<vec.size();++i)
+  for(auto i=from;i<static_cast<int64_t>(vec.size());++i)
     vec[i] = lut_missing_value;
 }
 
@@ -42,7 +42,7 @@ template<bool inputs_2_merged_LUT_is_input_ordered, bool merged_2_inputs_LUT_is_
 void LUTBase<inputs_2_merged_LUT_is_input_ordered, merged_2_inputs_LUT_is_input_ordered>::resize_and_reset_vector(vector<int64_t>& vec, int64_t new_size)
 {
   auto old_size = vec.size();
-  if(new_size > old_size)
+  if(new_size > static_cast<int64_t>(old_size))
   {
     vec.resize(new_size);
     reset_vector(vec, old_size);
@@ -54,7 +54,7 @@ void LUTBase<inputs_2_merged_LUT_is_input_ordered, merged_2_inputs_LUT_is_input_
 (vector<vector<int64_t>>& lut, int64_t new_lut_size, int64_t new_vector_size, int64_t& numRowsVar, int64_t& numColsVar)
 {
   auto old_lut_size = lut.size();
-  if(new_lut_size > old_lut_size)
+  if(new_lut_size > static_cast<int64_t>(old_lut_size))
   {
     lut.resize(new_lut_size);
     numRowsVar = new_lut_size;
@@ -62,14 +62,14 @@ void LUTBase<inputs_2_merged_LUT_is_input_ordered, merged_2_inputs_LUT_is_input_
   auto old_vector_size = (lut.size() > 0u) ? lut[0].size() : 0u;
   //Begin resizing of vectors at start_idx
   auto start_idx = old_lut_size;
-  if(new_vector_size > old_vector_size)     //every vector needs to be resized
+  if(new_vector_size > static_cast<int64_t>(old_vector_size))     //every vector needs to be resized
   {
     start_idx = 0u;
     numColsVar = new_vector_size;
   }
   else
     new_vector_size = old_vector_size;      //new vector size is smaller, don't bother reducing the size of existing rows
-  for(auto i=start_idx;i<new_lut_size;++i)
+  for(auto i=start_idx;static_cast<int64_t>(i)<new_lut_size;++i)
     resize_and_reset_vector(lut[i], new_vector_size);
 }
 //explicit initialization to avoid link errors
